+++ src/OpenTelemetry.Api/Internal/SpanHelper.cs
@@ -19,7 +19,7 @@ namespace OpenTelemetry.Trace
     /// <summary>
     /// A collection of helper methods to be used when building spans.
     /// </summary>
-    public static class SpanHelper
+    internal static class SpanHelper
     {
         /// <summary>
         /// Helper method that populates span properties from http status code according
@@ -29,14 +29,12 @@ namespace OpenTelemetry.Trace
         /// <returns>Resolved span <see cref="Status"/> for the Http status code.</returns>
         public static Status ResolveSpanStatusForHttpStatusCode(int httpStatusCode)
         {
-            var status = Status.Error;
-
             if (httpStatusCode >= 100 && httpStatusCode <= 399)
             {
-                status = Status.Unset;
+                return Status.Unset;
             }
 
-            return status;
+            return Status.Error;
         }
     }
 }