+++ src/OpenTelemetry.Instrumentation.Http/Implementation/HttpWebRequestActivitySource.netfx.cs
@@ -122,10 +122,7 @@ namespace OpenTelemetry.Instrumentation.Http.Implementation
             {
                 activity.SetTag(SemanticConventions.AttributeHttpStatusCode, (int)response.StatusCode);
 
-                activity.SetStatus(
-                    SpanHelper
-                        .ResolveSpanStatusForHttpStatusCode((int)response.StatusCode)
-                        .WithDescription(response.StatusDescription));
+                activity.SetStatus(SpanHelper.ResolveSpanStatusForHttpStatusCode((int)response.StatusCode));
 
                 try
                 {
@@ -153,7 +150,7 @@ namespace OpenTelemetry.Instrumentation.Http.Implementation
                 {
                     activity.SetTag(SemanticConventions.AttributeHttpStatusCode, (int)response.StatusCode);
 
-                    status = SpanHelper.ResolveSpanStatusForHttpStatusCode((int)response.StatusCode).WithDescription(response.StatusDescription);
+                    status = SpanHelper.ResolveSpanStatusForHttpStatusCode((int)response.StatusCode);
                 }
                 else
                 {
@@ -204,9 +201,20 @@ namespace OpenTelemetry.Instrumentation.Http.Implementation
 
         private static void ProcessRequest(HttpWebRequest request)
         {
-            if (!WebRequestActivitySource.HasListeners() || IsRequestInstrumented(request) || !Options.EventFilter(request))
+            if (!WebRequestActivitySource.HasListeners() || !Options.EventFilter(request))
             {
-                // No subscribers to the ActivitySource or this request was instrumented by previous
+                // No subscribers to the ActivitySource or User provider Filter is
+                // filtering this request.
+                // Propagation must still be done in such cases, to allow
+                // downstream services to continue from parent context, if any.
+                // Eg: Parent could be the Asp.Net activity.
+                InstrumentRequest(request, Activity.Current?.Context ?? default);
+                return;
+            }
+
+            if (IsRequestInstrumented(request))
+            {
+                // This request was instrumented by previous
                 // ProcessRequest, such is the case with redirect responses where the same request is sent again.
                 return;
             }
@@ -214,6 +222,9 @@ namespace OpenTelemetry.Instrumentation.Http.Implementation
             var activity = WebRequestActivitySource.StartActivity(ActivityName, ActivityKind.Client);
             var activityContext = Activity.Current?.Context ?? default;
 
+            // Propagation must still be done in all cases, to allow
+            // downstream services to continue from parent context, if any.
+            // Eg: Parent could be the Asp.Net activity.
             InstrumentRequest(request, activityContext);
             if (activity == null)
             {