+++ test/OpenTelemetry.Exporter.Jaeger.Tests/JaegerExporterTests.cs
@@ -28,6 +28,8 @@ namespace OpenTelemetry.Exporter.Jaeger.Tests
 {
     public class JaegerExporterTests
     {
+        private const string DefaultServiceName = "OpenTelemetry Exporter";
+
         [Fact]
         public void JaegerExporter_BadArgs()
         {
@@ -38,10 +40,7 @@ namespace OpenTelemetry.Exporter.Jaeger.Tests
         [Fact]
         public void JaegerTraceExporter_ctor_NullServiceNameAllowed()
         {
-            using var jaegerTraceExporter = new JaegerExporter(new JaegerExporterOptions
-            {
-                ServiceName = null,
-            });
+            using var jaegerTraceExporter = new JaegerExporter(new JaegerExporterOptions());
             Assert.NotNull(jaegerTraceExporter);
         }
 
@@ -57,11 +56,11 @@ namespace OpenTelemetry.Exporter.Jaeger.Tests
 
             Assert.Equal("TestService", process.ServiceName);
 
-            jaegerTraceExporter.SetResource(Resources.Resources.CreateServiceResource("MyService"));
+            jaegerTraceExporter.SetResource(ResourceBuilder.CreateEmpty().AddService("MyService").Build());
 
             Assert.Equal("MyService", process.ServiceName);
 
-            jaegerTraceExporter.SetResource(Resources.Resources.CreateServiceResource("MyService", serviceNamespace: "MyNamespace"));
+            jaegerTraceExporter.SetResource(ResourceBuilder.CreateEmpty().AddService("MyService", "MyNamespace").Build());
 
             Assert.Equal("MyNamespace.MyService", process.ServiceName);
         }
@@ -72,10 +71,10 @@ namespace OpenTelemetry.Exporter.Jaeger.Tests
             using var jaegerTraceExporter = new JaegerExporter(new JaegerExporterOptions());
             var process = jaegerTraceExporter.Process;
 
-            jaegerTraceExporter.SetResource(new Resource(new Dictionary<string, object>
+            jaegerTraceExporter.SetResource(ResourceBuilder.CreateEmpty().AddAttributes(new Dictionary<string, object>
             {
                 ["Tag"] = "value",
-            }));
+            }).Build());
 
             Assert.NotNull(process.Tags);
             Assert.Single(process.Tags);
@@ -90,10 +89,10 @@ namespace OpenTelemetry.Exporter.Jaeger.Tests
 
             process.Tags = new Dictionary<string, JaegerTag> { ["Tag1"] = new KeyValuePair<string, object>("Tag1", "value1").ToJaegerTag() };
 
-            jaegerTraceExporter.SetResource(new Resource(new Dictionary<string, object>
+            jaegerTraceExporter.SetResource(ResourceBuilder.CreateEmpty().AddAttributes(new Dictionary<string, object>
             {
                 ["Tag2"] = "value2",
-            }));
+            }).Build());
 
             Assert.NotNull(process.Tags);
             Assert.Equal(2, process.Tags.Count);
@@ -107,11 +106,11 @@ namespace OpenTelemetry.Exporter.Jaeger.Tests
             using var jaegerTraceExporter = new JaegerExporter(new JaegerExporterOptions());
             var process = jaegerTraceExporter.Process;
 
-            jaegerTraceExporter.SetResource(new Resource(new Dictionary<string, object>
+            jaegerTraceExporter.SetResource(ResourceBuilder.CreateEmpty().AddAttributes(new Dictionary<string, object>
             {
-                [Resource.ServiceNameKey] = "servicename",
-                [Resource.ServiceNamespaceKey] = "servicenamespace",
-            }));
+                [ResourceSemanticConventions.AttributeServiceName] = "servicename",
+                [ResourceSemanticConventions.AttributeServiceNamespace] = "servicenamespace",
+            }).Build());
 
             Assert.Null(process.Tags);
         }
@@ -120,7 +119,7 @@ namespace OpenTelemetry.Exporter.Jaeger.Tests
         public void JaegerTraceExporter_BuildBatchesToTransmit_DefaultBatch()
         {
             // Arrange
-            using var jaegerExporter = new JaegerExporter(new JaegerExporterOptions { ServiceName = "TestService" });
+            using var jaegerExporter = new JaegerExporter(new JaegerExporterOptions());
             jaegerExporter.SetResource(Resource.Empty);
 
             // Act
@@ -132,7 +131,7 @@ namespace OpenTelemetry.Exporter.Jaeger.Tests
 
             // Assert
             Assert.Single(batches);
-            Assert.Equal("TestService", batches.First().Process.ServiceName);
+            Assert.Equal(DefaultServiceName, batches.First().Process.ServiceName);
             Assert.Equal(3, batches.First().Count);
         }
 
@@ -140,7 +139,7 @@ namespace OpenTelemetry.Exporter.Jaeger.Tests
         public void JaegerTraceExporter_BuildBatchesToTransmit_MultipleBatches()
         {
             // Arrange
-            using var jaegerExporter = new JaegerExporter(new JaegerExporterOptions { ServiceName = "TestService" });
+            using var jaegerExporter = new JaegerExporter(new JaegerExporterOptions());
             jaegerExporter.SetResource(Resource.Empty);
 
             // Act
@@ -158,7 +157,7 @@ namespace OpenTelemetry.Exporter.Jaeger.Tests
             // Assert
             Assert.Equal(2, batches.Count());
 
-            var primaryBatch = batches.Where(b => b.Process.ServiceName == "TestService");
+            var primaryBatch = batches.Where(b => b.Process.ServiceName == DefaultServiceName);
             Assert.Single(primaryBatch);
             Assert.Equal(2, primaryBatch.First().Count);
 
@@ -171,7 +170,7 @@ namespace OpenTelemetry.Exporter.Jaeger.Tests
         public void JaegerTraceExporter_BuildBatchesToTransmit_FlushedBatch()
         {
             // Arrange
-            using var jaegerExporter = new JaegerExporter(new JaegerExporterOptions { ServiceName = "TestService", MaxPayloadSizeInBytes = 1500 });
+            using var jaegerExporter = new JaegerExporter(new JaegerExporterOptions { MaxPayloadSizeInBytes = 1500 });
             jaegerExporter.SetResource(Resource.Empty);
 
             // Act
@@ -183,7 +182,7 @@ namespace OpenTelemetry.Exporter.Jaeger.Tests
 
             // Assert
             Assert.Single(batches);
-            Assert.Equal("TestService", batches.First().Process.ServiceName);
+            Assert.Equal(DefaultServiceName, batches.First().Process.ServiceName);
             Assert.Equal(1, batches.First().Count);
         }