+++ src/OpenTelemetry.Exporter.ZPages/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.0.0-rc1.1
+
+Released 2020-Nov-17
+
 ## 0.8.0-beta.1
 
 Released 2020-Nov-5